+++ core/src/hir/mod.rs
@@ -2,16 +2,16 @@
 
 #![allow(unused_mut, unused_variables, dead_code)]
 mod defs;
-mod error;
 mod lifetimes;
+mod lowering;
 mod methods;
 mod paths;
 mod primitives;
 mod type_context;
 mod types;
 pub use defs::*;
-pub use error::*;
 pub use lifetimes::*;
+pub use lowering::*;
 pub use methods::*;
 pub use paths::*;
 pub use primitives::*;
